Symmetric Unification
Removed all the type specific unification protocols, users now only
have to implement IUnifyTerms. This eliminates a lot of pointless
boilerplate and introduces very little performance degradation.

Added INonStorable marker interface. PMap is now an instance of
INonStorable. It is unsound to be able to actually store PMap in the
substitution, trying to do so will now throw an exception. Update the
partial map tests to reflect this.

Symmetric unification also means there is no longer any general hook
to override things like map unification as was being done for PMap. So
that record unification can trump map unfication in the case of
(== map record), we've added IUnifyWithRecord and extended PMap to it as
an example.