Disequality that does not depend on unify.
Added IDisunifyTerms protocol. Added disunify which looks like unify
but instead returns a constraint set. Not yet sure if this approach is
general enough to implement nominal disequality. Extend the core type
to IDisunifyTerms.

No more calls to unify in !=c. Commented out normalize-store and
prefix-subsumes? as those are optimizations anyhow. != goal now calls
disunify and uses the :prefixc property of the constraint set to
construct the !=c constraint.