Deep constraints
Added `treec`. `treec` takes a term and a constraint and form to
display upon reification. `treec` will ensure that when a term becomes
ground that if it is a tree term that none of the sub terms violate
the constraint. Added an initial test.

Related to this change, fixed some constraints that didn't implement
`IConstraintId` which results in a space leak in the constraint store.