`treec` needs to support custom reification
It is like that people will want to reify usage of `treec` in whatever
way they please. Change `treec` so that `pform` can be function. If a
function then we pass the constraint, the reified constraint value,
the constraint substitution, as well as the full substitution if there
are values outside the constraint which should also be reified. Added
`creify` which can take a value, the constraint substitution and the
full substitution and do the right thing.