Optimize map-sum so we don't OOM when there are lots of options
Using conde for domains on lots of logic vars can result in
exponential memory usage due to conde's interleaving (which is itself
caused by conde wrapping its result in -inc). This commit changes
map-sum such that it should be equivalent to the old version without
the -inc. This should be a valid change since the unification goals
should not diverge.